refactor: move pl specifics to to pl::extra::stmt
