+++ prql-compiler/src/ast/pl/extra/mod.rs
@@ -1,1 +1,2 @@
 pub mod expr;
+pub mod stmt;
